DockerSource: Fix formula of download progress
When downloading image with multiple layers, the download progress
value of every following layer should not start from 0.

If we have 10 layers, downloading each of them should increase the
total download progress by 10%.

Assuming that the download and extraction are 50/50 of the total work.

Then, downloading each of 10 layers will increase the progress value
with 5% of the total work.

When all layers are downloaded the progress value should be 50%.

However, with the current formula the progress value of each layer
starts from 0%.
(E.g. when downloading 2nd layer of 10 the download progress starts
from 0% instead of 5%.)

This bug can be seen when downloading images with multiple layers of
large size.

Example:
    virt-bootstrap docker://rails /tmp/foo --status-only